moved the listing spesific template into its own file in the templates folder
